refactor: Fixed to pass target as an argument in validation options
